cleanup refactor
