+++ lib/behavior/onboard_new_user/callbackRegistry.js
@@ -1,0 +1,6 @@
+const store = {};
+module.exports = {
+  set: (id, handler) => store[id] = handler,
+  get: id => store[id],
+  clear: id => delete store[id]
+};
